# voice-journal-api-spec.yaml
openapi: 3.0.3
info:
  title: "Voice Journal API"
  version: "1.0.0"
  description: |
    Voice Journal은 사용자가 매일 밤 내일의 계획을 음성으로 녹음하고,
    아침에 자신의 목소리로 들으며 동기부여를 받는 개인 성장 플랫폼입니다.
    
    ## 주요 기능
    - 🎤 음성 파일 업로드 및 OpenAI Whisper를 통한 텍스트 변환
    - 📝 계획(plan)과 회고(reflection) 기록 관리
    - 🎯 목표 설정 및 달성률 추적
    - 📊 성장 패턴 분석 및 인사이트 제공
    
    ## 인증 방식
    이 API는 JWT 기반 Bearer 토큰 인증을 사용합니다.
    모든 보호된 엔드포인트는 Authorization 헤더에 'Bearer {token}' 형식으로 토큰을 포함해야 합니다.
    
    ## 응답 형식
    모든 응답은 JSON 형식이며, 에러 발생 시 일관된 ErrorResponse 스키마를 따릅니다.
    
    ## 파일 업로드
    음성 파일 업로드는 S3 Pre-signed URL을 통해 이루어지며, 
    지원되는 형식은 MP3, WAV, M4A, OGG (최대 25MB)입니다.
  termsOfService: "https://voicejournal.com/terms"
  contact:
    name: "Voice Journal 지원팀"
    url: "https://voicejournal.com/support"
    email: "support@voicejournal.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "https://api.voicejournal.com"
    description: "Production server"
  - url: "https://staging-api.voicejournal.com"
    description: "Staging server for testing"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT 토큰을 Authorization 헤더에 'Bearer {token}' 형식으로 포함해주세요."
  schemas:
    ErrorResponse:
      type: object
      description: "표준 에러 응답 형식"
      properties:
        code:
          type: integer
          description: "HTTP 상태 코드"
          example: 400
        message:
          type: string
          description: "사용자에게 표시할 에러 메시지"
          example: "잘못된 요청입니다."
        details:
          type: string
          description: "개발자를 위한 상세 에러 정보 (선택적)"
          example: "filename 파라미터가 누락되었습니다."
        timestamp:
          type: string
          format: date-time
          description: "에러 발생 시간 (ISO 8601)"
          example: "2024-01-15T10:30:00Z"
        requestId:
          type: string
          description: "요청 추적을 위한 고유 ID (선택적)"
          example: "req_abc123def456"
      required: [code, message, timestamp]
...
    VoiceEntry:
      type: object
      description: "음성 기록 정보"
      properties:
        id:
          type: string
          description: "기록 고유 식별자"
          example: "entry_789012345"
        userId:
          type: string
          description: "소유자 사용자 ID"
          example: "user_123456789"
        type:
          type: string
          enum: [plan, reflection]
          description: "기록 타입 - plan: 계획, reflection: 회고"
          example: "plan"
        date:
          type: string
          format: date
          description: "기록 날짜 (YYYY-MM-DD)"
          example: "2024-01-15"
        time:
          type: string
          pattern: "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: "기록 시간 (HH:MM, 24시간 형식)"
          example: "21:30"
        duration:
          type: string
          pattern: "^[0-9]+:[0-5][0-9]$"
          description: "음성 길이 (MM:SS)"
          example: "2:45"
        text:
          type: string
          description: "AI로 변환된 텍스트 내용"
          example: "내일은 아침 7시에 일어나서 운동을 30분 하고, 오전에는 프로젝트 기획서를 완성하겠습니다. 오후에는 독서 시간을 가지고, 저녁에는 오늘 하루를 돌아보는 시간을 갖겠습니다."
        audioUrl:
          type: string
          format: uri
          description: "음성 파일 접근 URL"
          example: "https://cdn.voicejournal.com/audio/user123/entry789012.mp3"
        fileName:
          type: string
          description: "업로드된 원본 파일명"
          example: "morning_plan.mp3"
        fileSize:
          type: string
          description: "파일 크기 (사람이 읽기 쉬운 형식)"
          example: "2.1MB"
        fileSizeBytes:
          type: integer
          description: "파일 크기 (바이트)"
          example: 2097152
        language:
          type: string
          description: "AI가 감지한 언어 코드"
          example: "ko"
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: "AI 변환 신뢰도 (0-1)"
          example: 0.95
        summary:
          type: string
          description: "AI가 생성한 내용 요약 (선택적)"
          example: "내일 아침 운동과 프로젝트 기획서 완성을 목표로 설정"
        keywords:
          type: array
          items:
            type: string
          description: "AI가 추출한 키워드 목록"
          example: ["운동", "프로젝트", "기획서", "독서"]
        completed:
          type: boolean
          description: "완료 여부 (계획의 경우)"
          example: false
        mood:
          type: string
          enum: [positive, neutral, negative]
          description: "AI가 분석한 감정 상태 (선택적)"
          example: "positive"
        createdAt:
          type: string
          format: date-time
          description: "생성 시간"
          example: "2024-01-15T21:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: "마지막 수정 시간"
          example: "2024-01-15T21:30:00Z"
      required: [id, userId, type, date, time, duration, text, audioUrl, createdAt]
...
    PaginationInfo:
      type: object
      description: "페이지네이션 정보"
      properties:
        page:
          type: integer
          minimum: 1
          description: "현재 페이지 번호"
          example: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
          description: "페이지당 항목 수"
          example: 10
        total:
          type: integer
          minimum: 0
          description: "전체 항목 수"
          example: 45
        totalPages:
          type: integer
          minimum: 0
          description: "전체 페이지 수"
          example: 5
        hasNext:
          type: boolean
          description: "다음 페이지 존재 여부"
          example: true
        hasPrev:
          type: boolean
          description: "이전 페이지 존재 여부"
          example: false
        nextPage:
          type: integer
          description: "다음 페이지 번호 (있는 경우)"
          example: 2
        prevPage:
          type: integer
          description: "이전 페이지 번호 (있는 경우)"
          example: null
      required: [page, size, total, totalPages, hasNext, hasPrev]
    PaginatedVoiceEntries:
      type: object
      description: "페이지네이션된 음성 기록 목록"
      properties:
        data:
          type: array
          description: "음성 기록 배열"
          items:
            $ref: '#/components/schemas/VoiceEntry'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        filters:
          type: object
          description: "적용된 필터 정보"
          properties:
            type:
              type: string
              enum: [plan, reflection]
              example: "plan"
            startDate:
              type: string
              format: date
              example: "2024-01-01"
            endDate:
              type: string
              format: date
              example: "2024-01-31"
            sort:
              type: string
              enum: [newest, oldest]
              example: "newest"
      required: [data, pagination]
...
paths:
...
  /api/v1/voice-entries:
    post:
      tags: ["Voice Entries"]
      summary: "음성 기록 생성"
      description: |
        새로운 음성 기록을 생성합니다.
        
        - 계획(plan) 또는 회고(reflection) 타입 지정
        - AI 변환된 텍스트와 원본 음성 파일 연결
        - 자동 키워드 추출 및 감정 분석
        - 무료 플랜 사용량 제한 적용
      operationId: "createVoiceEntry"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [plan, reflection]
                  description: "기록 타입"
                  example: "plan"
                text:
                  type: string
                  minLength: 1
                  description: "변환된 텍스트"
                  example: "내일은 아침 7시에 일어나서 운동을 30분 하겠습니다."
                audioUrl:
                  type: string
                  format: uri
                  description: "음성 파일 URL"
                  example: "https://cdn.voicejournal.com/audio/user123/entry789.mp3"
                fileName:
                  type: string
                  description: "원본 파일명"
                  example: "morning_plan.mp3"
                fileSize:
                  type: string
                  description: "파일 크기"
                  example: "2.1MB"
                fileSizeBytes:
                  type: integer
                  description: "파일 크기 (바이트)"
                  example: 2097152
                duration:
                  type: string
                  description: "음성 길이"
                  example: "2:45"
                language:
                  type: string
                  description: "언어 코드"
                  example: "ko"
                confidence:
                  type: number
                  format: float
                  description: "변환 신뢰도"
                  example: 0.95
                date:
                  type: string
                  format: date
                  description: "기록 날짜 (선택적, 기본값: 오늘)"
                  example: "2024-01-15"
              required: [type, text, audioUrl]
      responses:
        '201':
          description: "기록 생성 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceEntry'
        '400':
          description: "잘못된 요청"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: "무료 플랜 사용량 초과"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: "요청 한도 초과"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "서버 내부 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: ["Voice Entries"]
      summary: "음성 기록 목록 조회"
      description: |
        사용자의 음성 기록 목록을 페이지네이션으로 조회합니다.
        
        - 타입별 필터링 (계획/회고)
        - 날짜 범위 필터링
        - 정렬 옵션 (최신순/오래된순)
        - 검색 기능 (텍스트 내용 기반)
      operationId: "getVoiceEntries"
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: "페이지 번호"
          example: 1
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: "페이지 크기"
          example: 10
        - in: query
          name: type
          schema:
            type: string
            enum: [plan, reflection]
          description: "기록 타입 필터"
          example: "plan"
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: "시작 날짜 필터 (YYYY-MM-DD)"
          example: "2024-01-01"
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: "종료 날짜 필터 (YYYY-MM-DD)"
          example: "2024-01-31"
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
          description: "정렬 순서"
          example: "newest"
        - in: query
          name: search
          schema:
            type: string
            maxLength: 100
          description: "텍스트 내용 검색"
          example: "운동"
        - in: query
          name: completed
          schema:
            type: boolean
          description: "완료 여부 필터 (계획의 경우)"
          example: false
      responses:
        '200':
          description: "목록 조회 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVoiceEntries'
        '400':
          description: "잘못된 요청 (잘못된 날짜 형식 등)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: "요청 한도 초과"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "서버 내부 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    options:
      tags: ["Voice Entries"]
      summary: "CORS preflight for voice entries"
      description: "CORS preflight 요청을 처리합니다."
      operationId: "voiceEntriesOptions"
      responses:
        '204':
          description: "CORS preflight 성공"
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: "*"
            Access-Control-Allow-Methods:
              schema:
                type: string
                example: "GET, POST, OPTIONS"
            Access-Control-Allow-Headers:
              schema:
                type: string
                example: "Content-Type, Authorization"
  /api/v1/voice-entries/{entryId}:
    get:
      tags: ["Voice Entries"]
      summary: "음성 기록 상세 조회"
      description: |
        특정 음성 기록의 상세 정보를 조회합니다.
        
        - 전체 텍스트 내용
        - 음성 파일 정보
        - AI 분석 결과 (키워드, 감정 등)
        - 세그먼트별 타임스탬프 (있는 경우)
      operationId: "getVoiceEntry"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: entryId
          required: true
          schema:
            type: string
            pattern: "^entry_[a-zA-Z0-9]+$"
          description: "기록 고유 ID"
          example: "entry_789012345"
      responses:
        '200':
          description: "상세 조회 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceEntry'
        '400':
          description: "잘못된 요청 (잘못된 ID 형식)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "기록을 찾을 수 없음"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: "요청 한도 초과"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "서버 내부 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: ["Voice Entries"]
      summary: "음성 기록 수정"
      description: |
        기존 음성 기록을 수정합니다.
        
        - 텍스트 내용 수정 가능
        - 완료 상태 변경 가능
        - 음성 파일은 수정 불가
        - 수정 시간 자동 업데이트
      operationId: "updateVoiceEntry"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: entryId
          required: true
          schema:
            type: string
            pattern: "^entry_[a-zA-Z0-9]+$"
          description: "기록 고유 ID"
          example: "entry_789012345"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  minLength: 1
                  description: "수정된 텍스트"
                  example: "내일은 아침 7시에 일어나서 운동을 1시간 하겠습니다."
                completed:
                  type: boolean
                  description: "완료 여부"
                  example: true
                summary:
                  type: string
                  description: "사용자 정의 요약"
                  example: "내일 아침 운동 계획"
              minProperties: 1
      responses:
        '200':
          description: "수정 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceEntry'
        '400':
          description: "잘못된 요청"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "기록을 찾을 수 없음"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: "요청 한도 초과"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "서버 내부 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: ["Voice Entries"]
      summary: "음성 기록 삭제"
      description: |
        특정 음성 기록을 삭제합니다.
        
        - 기록과 연결된 음성 파일도 함께 삭제
        - 삭제된 데이터는 복구 불가
        - 소프트 삭제 (30일 후 완전 삭제)
      operationId: "deleteVoiceEntry"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: entryId
          required: true
          schema:
            type: string
            pattern: "^entry_[a-zA-Z0-9]+$"
          description: "기록 고유 ID"
          example: "entry_789012345"
      responses:
        '204':
          description: "삭제 성공"
        '400':
          description: "잘못된 요청"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "기록을 찾을 수 없음"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: "요청 한도 초과"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "서버 내부 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    options:
      tags: ["Voice Entries"]
      summary: "CORS preflight for voice entry operations"
      description: "CORS preflight 요청을 처리합니다."
      operationId: "voiceEntryOptions"
      responses:
        '204':
          description: "CORS preflight 성공"
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: "*"
            Access-Control-Allow-Methods:
              schema:
                type: string
                example: "GET, PUT, DELETE, OPTIONS"
            Access-Control-Allow-Headers:
              schema:
                type: string
                example: "Content-Type, Authorization"

  # Goals
...